Move logging to top level, rename submission.log
