+++ submission/alignment/system3.py
@@ -6,7 +6,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import scipy.stats
 
-from ..loggers import logger
+from ..log import logger
 from ..parsers.system3_log_parser import System3LogParser
 from .system1 import UnAlignableEEGException
