+++ submission/detection/artifact_detection.py
@@ -1,7 +1,7 @@
 import os
 import glob
 import numpy as np
-from ..loggers import logger
+from ..log import logger
 from ..helpers.butter_filt import butter_filt
