--- submission/loggers/__init__.py
@@ -1,3 +1,0 @@
-from .logger import Logger
-
-logger = Logger()
