+++ submission/parsers/base_log_parser.py
@@ -7,7 +7,7 @@ import sqlite3
 import numpy as np
 import pandas as pd
 
-from ..loggers import logger
+from ..log import logger
 from ..readers.eeg_reader import read_jacksheet
 from ..viewers.view_recarray import pformat_rec, to_dict, from_dict
