+++ submission/parsers/system3_log_parser.py
@@ -5,7 +5,7 @@ import numpy as np
 from .base_log_parser import BaseLogParser,BaseSys3LogParser
 from .fr_log_parser import FRSessionLogParser
 from .electrode_config_parser import ElectrodeConfig
-from ..loggers import logger
+from ..log import logger
 
 
 class System3LogParser(object):
